pyproject.toml
[project]
name = "adk2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Marcello Steiner", email = "msteiner@google.com" }]
requires-python = ">=3.14"
dependencies = [
  "a2a-sdk==0.3.25",
  "a2ui-agent-sdk>=0.2.1",
  "google-adk==2.0.0a3",
  "google-cloud-aiplatform[agent-engines]>=1.148.1",
  "jinja2>=3.1.3",
  "loguru>=0.7.3",
  "pydantic>=2.13.3",
  "rapidfuzz>=3.14.5",
]

[project.scripts]
adk2 = "adk2.main:serve"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "data",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
  "**/dist",
  "test_data",
  "tests",
]
# Lint and format also notebooks.
extend-include = ["**/*.ipynb"]
# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.14
target-version = "py314"

[tool.ruff.format]

preview = true

# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doc tests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = true

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"


[tool.ruff.lint]
preview = true
select = [
  "A",
  "AIR",
  "ANN",
  "ARG",
  "B",
  "BLE",
  "C4",
  "C90",
  "COM",
  "D",
  "DTZ",
  "E",
  "ERA",
  "F",
  "FBT",
  "FLY",
  "FURB",
  "G",
  "I",
  "INP",
  "INT",
  "ISC",
  "LOG",
  "N",
  "NPY",
  "PD",
  "PERF",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "Q",
  "RET",
  "RUF",
  "S",
  "SIM",
  "SLF",
  "T20",
  "TC",
  "TID",
  "TRY",
  "UP",
  "W",
]
ignore = ["COM812", "ISC001", "S101"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"**/*.ipynb" = ["ANN", "D", "A004", "E402", "PTH", "E501", "T20"]

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.basedpyright]
include = ["src"]
typeCheckingMode = "off"
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "src/experimental",
  "src/typestubs",
  ".venv/",
]
ignore = ["src/oldstuff"]

reportMissingImports = true
reportMissingTypeStubs = false
reportCallIssue = false

pythonVersion = "3.14"

[dependency-groups]
dev = ["pytest>=8.0.0", "ruff>=0.15.11", "ty>=0.0.32"]

[tool.ty.environment]
python-version = "3.14"
root = [".src/"]

[[tool.ty.overrides]]
include = ["src/", "tests/"]
exclude = [
  "**/node_modules",
  "**/__pycache__",
  "src/experimental",
  "src/typestubs",
  ".venv/",
]

[tool.ty.rules]
# Rules can be set to "error", "warn", or "ignore"
# unresolved-import = "ignore"


[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
